feat(oidc): support Keycloak authorization
VEN-1583.

Refactor the authService to support Helsinki-Profile Keycloak
authorization. Add AppConfig.ts for consistency between the services of
city of Helsinki and to simplify the usage and configuration.

Tunnistamo and Keycloak can both be used and the change between them can
be easily done. Tunnistamo is good for local usage and at the moment, it
is also the only option for production use. However, Tunnistamo will be
replaced with Helsinki Profile Keycloak service soon.